Adding exceptions

+++ src/edu/oswego/csc436/data/RightEncoder.java
@@ -20,6 +20,6 @@ public class RightEncoder extends Encoder {
 
   @Override
   public void writeToEncoder(Motion motion, int value) {
-    motion.setRightSpeed(value);
+    motion.setRightSpeed((int) value);
   }
 }
